requirements.txt
kafka-python==2.2.15
psycopg2-binary==2.9.10
